// src/common/common-module/src/enums/collection.enum.ts
export enum CollectionStatus {
  IPC = 1,
  CLOSED = 'Closed',
  PART_PAYMENT = 'Part Payment',
  SETTLEMENT = 'Settlement',
  DISBURSED = 'Disbursed',
  EMI_CLOSED = 'EMI Closed',
  APPROVED = 'Approved',
  PAYMENT_REJECTED = 'Payment Rejected',
  APPROVAL_WAITING = 'Approval Waiting',
  LEAD_ALLOCATED = 'Lead Allocated',
  APPROVAL_WAITING_REFUNDED = 'Approval Waiting (refunded amount)',
  APPROVED_REFUNDED = 'Approved-refunded',
  WAIVE_OFF = 'Waive Off',
  REJECTED = 'Rejected',
}

export enum CollectedMode {
  ACCOUNT = 'Account',
  UPI = 'UPI',
  CASH = 'Cash',
  PAYTM = 'PayTM',
  E_MANDATE = 'E-Mandate',
  PAYMENT_GATEWAY = 'Payment Getway',
  HYPTO_BANK_VIRTUAL_AC = 'HYPTO Bank- Virtual A/c',
  RAZORPAY_BANK = 'Razorpay Bank',
  RAZORPAY_PAYMENT_GATEWAY = 'RAZORPAY PAYMENT GATEWAY',
  GOOGLE_PAY = 'GOOGLE PAY',
  CASHFREE_PAYMENT_GATEWAY = 'CASHFREE PAYMENT GATEWAY',
  ICICI_BANK_1435 = 'ICICI BANK-1435',
  ICICI_BANK_7788 = 'ICICI BANK-7788',
  CUB = 'CUB',
  IDBI_BANK_4213 = 'IDBI Bank-4213',
  SOUTH_INDIAN_4754 = 'SOUTH INDIAN-4754',
  CSB_BANK_5000 = 'CSB BANK-5000',
  INDUSIND_BANK_9999 = 'INDUSIND BANK-9999',
  BHARAT_SONI_PAYTM = 'BHARAT SONI PAYTM',
  DISCOUNT = 'DISCOUNT',
  WAIVE_OFF = 'Waive Off',
  PAYMENT_PENDING = 'Payment Pending',
  PAY_U = 'PayU',
  PAY_U_BILL_PAYMENT = 'Payu Bill Payment',
  PAYOUT = 'Payout',
  MANUAL = 'manual',
}

export enum CollectionType {
  APPROVAL_PENDING = 'Approval Pending',
  APPROVED = 'Approved',
  REJECTED = 'Rejected',
  ONLINE_PAYMENTS = 'Online Payments',
}